Bit more refactoring for sake of testing   - As you, dear reviewer, may be aware, there are complications     involved when trying to mock out a method from the same module as     your system under test     - Random Medium article explaining this:     - https://medium.com/welldone-software/jest-how-to-mock-a-function-call-inside-a-module-21c05c57a39f   - Here, I chose to simply break apart the `smartyStreets` module into     two, so that our imported and mocked functions will refer to the     same functions being executed by the code
