+++ frontend/src/app/utils/smartyStreets.ts
@@ -1,5 +1,6 @@
-import { core, usStreet, usZipcode } from "smartystreets-javascript-sdk";
+import { usStreet, usZipcode } from "smartystreets-javascript-sdk";
 
+import { getZipCodeClient, getClient } from "./smartyStreetsClients";
 import { toLowerStripWhitespace } from "./text";
 
 // cf. https://github.com/DefinitelyTyped/DefinitelyTyped/blob/11436c5a19fc6aabfd6b8d93b37dac38b4ab9bc2/types/smartystreets-javascript-sdk/index.d.ts#L625
@@ -7,13 +8,6 @@ export type ZipCodeResult = RequiredExceptFor<
   usZipcode.Result,
   "status" | "reason"
 >;
-
-export class SmartyStreetsError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "SmartyStreetsError";
-  }
-}
 
 const getLookup = (address: Address) => {
   const lookup = new usStreet.Lookup();
@@ -110,24 +104,4 @@ export function isValidZipCodeForState(
   }
 
   return result.zipcodes[0].stateAbbreviation === state;
-}
-
-export function buildClient(builder: Function) {
-  if (process.env.REACT_APP_SMARTY_STREETS_KEY === undefined) {
-    throw new SmartyStreetsError("Missing REACT_APP_SMARTY_STREETS_KEY");
-  }
-
-  const credentials = new core.SharedCredentials(
-    process.env.REACT_APP_SMARTY_STREETS_KEY
-  );
-
-  return builder(credentials);
-}
-
-export function getClient() {
-  return buildClient(core.buildClient.usStreet);
-}
-
-export function getZipCodeClient() {
-  return buildClient(core.buildClient.usZipcode);
 }
